refactor(scripts): centralize schema and add DB ownership fix
Address PR review feedback:

- scripts/create-schema.mjs: New shared schema module prevents duplication
- scripts/init-minimal-db.mjs: Import shared schema, add chown for DB file
- scripts/build-db.mjs: Remove inline schema, import from shared module
- Dockerfile: Add DB_FILE_UID/GID env vars, comment COPY ./data
- Dockerfile.dev: Add DB_FILE_UID/GID env vars

Schema is now defined in one place and used by both build tools.
DB file ownership is set to match runtime user (1001:1001).
